feat(web): enhance app and agent card interactions
