+++ web/app/components/apps/__tests__/app-card.spec.tsx
@@ -692,12 +692,36 @@ describe('AppCard', () => {
       expect(screen.getByRole('listitem')).toContainElement(cardLink)
     })
 
-    it('should expose a visible focus ring on the card link', () => {
+    it('should expose the card focus ring from the visual surface', () => {
       render(<AppCard app={mockApp} />)
       const cardLink = screen.getByRole('link', { name: 'Test App' })
+      const card = screen.getByRole('listitem')
 
-      expect(cardLink).toHaveClass('focus-visible:ring-2')
-      expect(cardLink).toHaveClass('focus-visible:ring-state-accent-solid')
+      expect(cardLink).toHaveClass('outline-hidden')
+      expect(cardLink.parentElement).toBe(card)
+      expect(card).toHaveClass('isolate')
+      expect(card).toHaveClass('after:z-1')
+      expect(card).toHaveClass('has-[>a:focus-visible]:after:inset-ring-2')
+      expect(card).toHaveClass('has-[>a:focus-visible]:after:inset-ring-state-accent-solid')
+    })
+
+    it('should keep card navigation and actions as sibling focus targets', async () => {
+      const user = userEvent.setup()
+      render(<AppCard app={mockApp} />)
+
+      const cardLink = screen.getByRole('link', { name: 'Test App' })
+      const starToggle = screen.getByRole('button', { name: 'app.studio.starApp: Test App' })
+      const operationsTrigger = getOperationsTrigger()
+
+      expect(cardLink).not.toContainElement(starToggle)
+      expect(cardLink).not.toContainElement(operationsTrigger)
+
+      await user.tab()
+      expect(cardLink).toHaveFocus()
+      await user.tab()
+      expect(starToggle).toHaveFocus()
+      await user.tab()
+      expect(operationsTrigger).toHaveFocus()
     })
 
     it('should star the app from the card action without navigating', async () => {
@@ -744,12 +768,14 @@ describe('AppCard', () => {
       render(<AppCard app={mockApp} />)
       const operationsTrigger = getOperationsTrigger()
       const operationsTriggerWrapper = operationsTrigger.closest('.absolute')
+      const actionBar = operationsTriggerWrapper?.firstElementChild
 
-      expect(operationsTriggerWrapper).toHaveClass('top-2')
-      expect(operationsTriggerWrapper).toHaveClass('right-2')
-      expect(operationsTriggerWrapper).toHaveClass('group-focus-within:pointer-events-auto')
+      expect(operationsTriggerWrapper).toHaveClass('top-[-0.5px]')
+      expect(operationsTriggerWrapper).toHaveClass('right-[-0.5px]')
+      expect(operationsTriggerWrapper).toHaveClass('h-16')
+      expect(operationsTriggerWrapper).toHaveClass('w-30')
       expect(operationsTriggerWrapper).toHaveClass('group-focus-within:opacity-100')
-      expect(operationsTriggerWrapper).not.toHaveClass('w-[120px]')
+      expect(actionBar).toHaveClass('group-focus-within:pointer-events-auto')
       expect(operationsTrigger).toHaveClass('focus-visible:ring-2')
       expect(operationsTrigger).toHaveClass('focus-visible:ring-state-accent-solid')
     })
@@ -764,6 +790,32 @@ describe('AppCard', () => {
         expect(screen.getByText('app.editApp')).toBeInTheDocument()
       })
       expect(mockPush).not.toHaveBeenCalled()
+    })
+
+    it('should expose the same operations from the card context menu', async () => {
+      const user = userEvent.setup()
+      render(<AppCard app={mockApp} />)
+
+      await user.pointer({
+        target: screen.getByRole('link', { name: 'Test App' }),
+        keys: '[MouseRight]',
+      })
+
+      expect(await screen.findByRole('menuitem', { name: 'app.editApp' })).toBeInTheDocument()
+      expect(screen.getByRole('menuitem', { name: 'app.duplicate' })).toBeInTheDocument()
+      expect(screen.getByRole('menuitem', { name: 'app.export' })).toBeInTheDocument()
+    })
+
+    it('should keep card actions outside the card context menu trigger', async () => {
+      const user = userEvent.setup()
+      render(<AppCard app={mockApp} />)
+
+      await user.pointer({
+        target: screen.getByRole('button', { name: 'app.studio.starApp: Test App' }),
+        keys: '[MouseRight]',
+      })
+
+      expect(screen.queryByRole('menuitem', { name: 'app.editApp' })).not.toBeInTheDocument()
     })
 
     it('should show duplicate option when dropdown menu is opened', async () => {
@@ -1203,7 +1255,10 @@ describe('AppCard', () => {
       render(<AppCard app={mockApp} />)
 
       const trigger = screen.getByRole('button', { name: 'common.operation.exporting' })
+      const actionBar = trigger.closest('.absolute')?.firstElementChild
+
       expect(trigger).toBeDisabled()
+      expect(actionBar).toHaveClass('pointer-events-auto')
     })
   })
