+++ web/features/agent-v2/roster/components/agent-roster-list.tsx
@@ -5,6 +5,13 @@ import { zAgentIconType } from '@dify/contracts/api/console/agent/zod.gen'
 import { Button } from '@langgenius/dify-ui/button'
 import { cn } from '@langgenius/dify-ui/cn'
 import {
+  ContextMenu,
+  ContextMenuContent,
+  ContextMenuItem,
+  ContextMenuSeparator,
+  ContextMenuTrigger,
+} from '@langgenius/dify-ui/context-menu'
+import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
@@ -105,9 +112,9 @@ function AgentRosterPlaceholderState({
   const { t } = useTranslation('common')
 
   return (
-    <section
+    <div
       aria-labelledby="agent-roster-placeholder-title"
-      className="relative col-span-full min-h-[calc(100vh-142px)] overflow-hidden"
+      className="relative min-h-[calc(100vh-142px)] overflow-hidden"
       role={role}
     >
       <div
@@ -141,22 +148,65 @@ function AgentRosterPlaceholderState({
           )}
         </div>
       </div>
-    </section>
+    </div>
+  )
+}
+
+type AgentCardActionMenuItemsProps = {
+  kind: 'context' | 'dropdown'
+  isExporting: boolean
+  onDelete: () => void
+  onDuplicate: () => void
+  onEdit: () => void
+  onExport: () => void
+}
+
+function AgentCardActionMenuItems({
+  kind,
+  isExporting,
+  onDelete,
+  onDuplicate,
+  onEdit,
+  onExport,
+}: AgentCardActionMenuItemsProps) {
+  const { t } = useTranslation('agentV2')
+  const { t: tCommon } = useTranslation('common')
+  const { t: tApp } = useTranslation('app')
+  const MenuItem = kind === 'context' ? ContextMenuItem : DropdownMenuItem
+  const MenuSeparator = kind === 'context' ? ContextMenuSeparator : DropdownMenuSeparator
+
+  return (
+    <>
+      <MenuItem className="gap-2" onClick={onEdit}>
+        <span aria-hidden className="i-ri-edit-line size-4 shrink-0 text-text-tertiary" />
+        <span>{t(($) => $['roster.editInfo'])}</span>
+      </MenuItem>
+      <MenuItem className="gap-2" onClick={onDuplicate}>
+        <span aria-hidden className="i-ri-file-copy-line size-4 shrink-0 text-text-tertiary" />
+        <span>{tCommon(($) => $['operation.duplicate'])}</span>
+      </MenuItem>
+      <MenuItem className="gap-2" disabled={isExporting} onClick={onExport}>
+        <span aria-hidden className="i-ri-download-line size-4 shrink-0 text-text-tertiary" />
+        <span>{tApp(($) => $.export)}</span>
+      </MenuItem>
+      <MenuSeparator />
+      <MenuItem variant="destructive" className="gap-2" onClick={onDelete}>
+        <span aria-hidden className="i-ri-delete-bin-line size-4 shrink-0" />
+        <span>{tCommon(($) => $['operation.delete'])}</span>
+      </MenuItem>
+    </>
   )
 }
 
 function AgentRosterItem({ agent }: { agent: AgentAppPartial }) {
   const { t } = useTranslation('agentV2')
-  const { t: tCommon } = useTranslation('common')
   const { t: tApp } = useTranslation('app')
   const { formatTime } = useTimestamp()
   const nameId = useId()
   const descriptionId = useId()
-  const [isEditOpen, setIsEditOpen] = useState(false)
+  const [activeDialog, setActiveDialog] = useState<'delete' | 'duplicate' | 'edit' | null>(null)
   const [editSessionKey, setEditSessionKey] = useState(0)
-  const [isDuplicateOpen, setIsDuplicateOpen] = useState(false)
   const [duplicateSessionKey, setDuplicateSessionKey] = useState(0)
-  const [isDeleteOpen, setIsDeleteOpen] = useState(false)
   const { exportAppDsl, isExporting } = useExportAppDsl()
   const updatedAt =
     agent.updated_at != null
@@ -175,12 +225,20 @@ function AgentRosterItem({ agent }: { agent: AgentAppPartial }) {
 
   const handleEditOpen = () => {
     setEditSessionKey((key) => key + 1)
-    setIsEditOpen(true)
+    setActiveDialog('edit')
   }
 
   const handleDuplicateOpen = () => {
     setDuplicateSessionKey((key) => key + 1)
-    setIsDuplicateOpen(true)
+    setActiveDialog('duplicate')
+  }
+
+  const handleDeleteOpen = () => {
+    setActiveDialog('delete')
+  }
+
+  const handleDialogOpenChange = (open: boolean) => {
+    if (!open) setActiveDialog(null)
   }
 
   const handleExport = () => {
@@ -196,49 +254,64 @@ function AgentRosterItem({ agent }: { agent: AgentAppPartial }) {
   }
 
   return (
-    <article
+    <li
       aria-labelledby={nameId}
-      className="group relative isolate col-span-1 h-36.5 min-w-0 overflow-hidden rounded-xl border-[0.5px] border-solid border-components-card-border bg-components-card-bg shadow-xs shadow-shadow-shadow-3 transition-shadow duration-200 ease-in-out after:pointer-events-none after:absolute after:inset-0 after:z-1 after:rounded-xl after:content-[''] focus-within:bg-components-card-bg-alt hover:bg-components-card-bg-alt hover:shadow-md hover:shadow-shadow-shadow-5 has-data-popup-open:bg-components-card-bg-alt has-[>a:focus-visible]:after:inset-ring-2 has-[>a:focus-visible]:after:inset-ring-state-accent-solid motion-reduce:transition-none [@media(hover:none)]:bg-components-card-bg-alt"
+      className="group relative isolate col-span-1 h-36.5 min-w-0 overflow-hidden rounded-xl border-[0.5px] border-solid border-components-card-border bg-components-card-bg shadow-xs shadow-shadow-shadow-3 transition-shadow duration-200 ease-in-out after:pointer-events-none after:absolute after:inset-0 after:z-1 after:rounded-xl after:content-[''] focus-within:bg-components-card-bg-alt hover:bg-components-card-bg-alt hover:shadow-md hover:shadow-shadow-shadow-5 has-data-popup-open:bg-components-card-bg-alt has-data-popup-open:shadow-md has-data-popup-open:shadow-shadow-shadow-5 has-[>a:focus-visible]:after:inset-ring-2 has-[>a:focus-visible]:after:inset-ring-state-accent-solid motion-reduce:transition-none [@media(hover:none)]:bg-components-card-bg-alt"
     >
-      <Link
-        href={`/agents/${agent.id}/configure`}
-        aria-labelledby={nameId}
-        aria-describedby={agent.description ? descriptionId : undefined}
-        className="flex h-full min-w-0 cursor-pointer touch-manipulation flex-col rounded-xl outline-hidden"
-      >
-        <div className="flex items-center gap-3 pt-3.5 pr-4 pb-2 pl-3.5">
-          <span aria-hidden className="shrink-0">
-            <AppIcon
-              size="xl"
-              rounded
-              iconType={iconType}
-              icon={agent.icon ?? undefined}
-              background={agent.icon_background}
-              imageUrl={imageUrl}
-            />
-          </span>
-          <div className="flex min-w-0 flex-1 flex-col gap-0.5 py-px">
-            <h2 id={nameId} className="truncate system-md-semibold text-text-secondary">
-              {agent.name}
-            </h2>
-            <p className="truncate system-xs-regular text-text-tertiary">{agent.role}</p>
-          </div>
-        </div>
-        <div className="px-4 py-1 system-xs-regular text-text-tertiary">
-          <div id={descriptionId} className="line-clamp-2 min-h-8">
-            {agent.description}
-          </div>
-        </div>
-        <div aria-hidden className="h-9 shrink-0" />
-        {isDraft && (
-          <div className="pointer-events-none absolute top-[-0.5px] right-0 flex h-5 items-start overflow-hidden">
-            <div className="h-5 w-3 bg-background-section-burn [clip-path:polygon(0_0,100%_0,100%_100%)]" />
-            <div className="flex h-5 items-center bg-background-section-burn pr-2 pl-0.5 system-2xs-medium-uppercase text-text-tertiary">
-              {t(($) => $['roster.usageStatus.draft'])}
-            </div>
-          </div>
-        )}
-      </Link>
+      <ContextMenu>
+        <ContextMenuTrigger
+          render={
+            <Link
+              href={`/agents/${agent.id}/configure`}
+              aria-labelledby={nameId}
+              aria-describedby={agent.description ? descriptionId : undefined}
+              className="flex h-full min-w-0 cursor-pointer touch-manipulation flex-col rounded-xl pb-9 outline-hidden"
+            >
+              <div className="flex items-center gap-3 pt-3.5 pr-4 pb-2 pl-3.5">
+                <span aria-hidden className="shrink-0">
+                  <AppIcon
+                    size="xl"
+                    rounded
+                    iconType={iconType}
+                    icon={agent.icon ?? undefined}
+                    background={agent.icon_background}
+                    imageUrl={imageUrl}
+                  />
+                </span>
+                <div className="flex min-w-0 flex-1 flex-col gap-0.5 py-px">
+                  <h2 id={nameId} className="truncate system-md-semibold text-text-secondary">
+                    {agent.name}
+                  </h2>
+                  <p className="truncate system-xs-regular text-text-tertiary">{agent.role}</p>
+                </div>
+              </div>
+              <div className="px-4 py-1 system-xs-regular text-text-tertiary">
+                <div id={descriptionId} className="line-clamp-2 min-h-8">
+                  {agent.description}
+                </div>
+              </div>
+              {isDraft && (
+                <div className="pointer-events-none absolute top-[-0.5px] right-0 flex h-5 items-start overflow-hidden">
+                  <div className="h-5 w-3 bg-background-section-burn [clip-path:polygon(0_0,100%_0,100%_100%)]" />
+                  <div className="flex h-5 items-center bg-background-section-burn pr-2 pl-0.5 system-2xs-medium-uppercase text-text-tertiary">
+                    {t(($) => $['roster.usageStatus.draft'])}
+                  </div>
+                </div>
+              )}
+            </Link>
+          }
+        />
+        <ContextMenuContent className="w-40">
+          <AgentCardActionMenuItems
+            kind="context"
+            isExporting={isExporting}
+            onEdit={handleEditOpen}
+            onDuplicate={handleDuplicateOpen}
+            onExport={handleExport}
+            onDelete={handleDeleteOpen}
+          />
+        </ContextMenuContent>
+      </ContextMenu>
       <div className="pointer-events-none absolute top-[-0.5px] right-[-0.5px] flex h-16 w-30 items-start justify-end bg-[linear-gradient(67deg,var(--color-components-card-bg-alt-transparent)_0%,var(--color-components-card-bg-alt)_75%)] p-2 opacity-0 group-focus-within:opacity-100 group-hover:opacity-100 has-data-popup-open:opacity-100 [@media(hover:none)]:opacity-100">
         <div className="pointer-events-none flex items-center overflow-hidden rounded-[10px] border-[0.5px] border-components-actionbar-border bg-components-actionbar-bg p-0.5 shadow-lg backdrop-blur-xs group-focus-within:pointer-events-auto group-hover:pointer-events-auto has-data-popup-open:pointer-events-auto [@media(hover:none)]:pointer-events-auto">
           <DropdownMenu modal={false}>
@@ -254,33 +327,14 @@ function AgentRosterItem({ agent }: { agent: AgentAppPartial }) {
               }
             />
             <DropdownMenuContent placement="bottom-end" sideOffset={4} className="w-40">
-              <DropdownMenuItem className="gap-2" onClick={handleEditOpen}>
-                <span aria-hidden className="i-ri-edit-line size-4 shrink-0 text-text-tertiary" />
-                <span>{t(($) => $['roster.editInfo'])}</span>
-              </DropdownMenuItem>
-              <DropdownMenuItem className="gap-2" onClick={handleDuplicateOpen}>
-                <span
-                  aria-hidden
-                  className="i-ri-file-copy-line size-4 shrink-0 text-text-tertiary"
-                />
-                <span>{tCommon(($) => $['operation.duplicate'])}</span>
-              </DropdownMenuItem>
-              <DropdownMenuItem className="gap-2" disabled={isExporting} onClick={handleExport}>
-                <span
-                  aria-hidden
-                  className="i-ri-download-line size-4 shrink-0 text-text-tertiary"
-                />
-                <span>{tApp(($) => $.export)}</span>
-              </DropdownMenuItem>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem
-                variant="destructive"
-                className="gap-2"
-                onClick={() => setIsDeleteOpen(true)}
-              >
-                <span aria-hidden className="i-ri-delete-bin-line size-4 shrink-0" />
-                <span>{tCommon(($) => $['operation.delete'])}</span>
-              </DropdownMenuItem>
+              <AgentCardActionMenuItems
+                kind="dropdown"
+                isExporting={isExporting}
+                onEdit={handleEditOpen}
+                onDuplicate={handleDuplicateOpen}
+                onExport={handleExport}
+                onDelete={handleDeleteOpen}
+              />
             </DropdownMenuContent>
           </DropdownMenu>
         </div>
@@ -316,24 +370,24 @@ function AgentRosterItem({ agent }: { agent: AgentAppPartial }) {
         </div>
       </div>
       <EditAgentDialog
+        key={editSessionKey}
         agent={agent}
-        formKey={editSessionKey}
-        open={isEditOpen}
-        onOpenChange={setIsEditOpen}
+        open={activeDialog === 'edit'}
+        onOpenChange={handleDialogOpenChange}
       />
       <DuplicateAgentDialog
+        key={duplicateSessionKey}
         agent={agent}
-        formKey={duplicateSessionKey}
-        open={isDuplicateOpen}
-        onOpenChange={setIsDuplicateOpen}
+        open={activeDialog === 'duplicate'}
+        onOpenChange={handleDialogOpenChange}
       />
       <DeleteAgentDialog
         agentId={agent.id}
         agentName={agent.name}
-        open={isDeleteOpen}
-        onOpenChange={setIsDeleteOpen}
+        open={activeDialog === 'delete'}
+        onOpenChange={handleDialogOpenChange}
       />
-    </article>
+    </li>
   )
 }
 
@@ -343,8 +397,12 @@ export function AgentRosterList({ label, state }: AgentRosterListProps) {
   const isBusy = state.status === 'pending' || (state.status === 'ready' && state.isFetching)
 
   return (
-    <section aria-label={label} className={AGENT_ROSTER_GRID_CLASS_NAME} aria-busy={isBusy}>
-      {state.status === 'pending' && <AgentRosterSkeleton />}
+    <section aria-label={label} aria-busy={isBusy}>
+      {state.status === 'pending' && (
+        <div className={AGENT_ROSTER_GRID_CLASS_NAME}>
+          <AgentRosterSkeleton />
+        </div>
+      )}
       {state.status === 'error' && (
         <AgentRosterPlaceholderState
           onRetry={state.onRetry}
@@ -363,10 +421,18 @@ export function AgentRosterList({ label, state }: AgentRosterListProps) {
         />
       )}
       {state.status === 'ready' &&
-        state.agents.map((agent) => <AgentRosterItem key={agent.id} agent={agent} />)}
+        state.agents.length > 0 && (
+          // Safari list semantics: https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/Properties/list-style#accessibility
+          // oxlint-disable-next-line jsx-a11y/no-redundant-roles -- Dify's preflight removes list markers.
+          <ul role="list" className={AGENT_ROSTER_GRID_CLASS_NAME}>
+            {state.agents.map((agent) => (
+              <AgentRosterItem key={agent.id} agent={agent} />
+            ))}
+          </ul>
+        )}
       {state.status === 'ready' && state.footer.status === 'error' && (
         <div
-          className="col-span-full flex items-center justify-center gap-3 pt-1 system-xs-regular text-text-destructive"
+          className="flex items-center justify-center gap-3 pt-1 system-xs-regular text-text-destructive"
           role="alert"
         >
           <span>{t(($) => $['roster.loadingError'])}</span>
@@ -376,7 +442,7 @@ export function AgentRosterList({ label, state }: AgentRosterListProps) {
         </div>
       )}
       {state.status === 'ready' && state.footer.status === 'load-more' && (
-        <div className="col-span-full flex justify-center pt-1">
+        <div className="flex justify-center pt-1">
           <Button loading={state.footer.isLoading} onClick={state.footer.onLoadMore}>
             {t(($) => $['roster.loadMore'])}
           </Button>
