+++ web/features/agent-v2/roster/components/edit-agent-dialog.tsx
@@ -25,7 +25,6 @@ import { AgentFormFields } from './agent-form-fields'
 
 type EditAgentDialogProps = {
   agent: AgentAppPartial
-  formKey: number
   open: boolean
   onOpenChange: (open: boolean) => void
 }
@@ -43,10 +42,9 @@ const applyIconPayload = (body: AgentAppUpdatePayload, icon: AgentIconSelection)
   body.icon_background = undefined
 }
 
-export function EditAgentDialog({ agent, formKey, open, onOpenChange }: EditAgentDialogProps) {
+export function EditAgentDialog({ agent, open, onOpenChange }: EditAgentDialogProps) {
   const { t } = useTranslation('agentV2')
   const { t: tCommon } = useTranslation('common')
-  const [renderedFormKey, setRenderedFormKey] = useState(formKey)
   const [name, setName] = useState(agent.name)
   const [description, setDescription] = useState(agent.description ?? '')
   const [role, setRole] = useState(agent.role ?? '')
@@ -56,24 +54,8 @@ export function EditAgentDialog({ agent, formKey, open, onOpenChange }: EditAgen
   )
   const updateAgentMutation = useMutation(consoleQuery.agent.byAgentId.put.mutationOptions())
 
-  if (formKey !== renderedFormKey) {
-    setRenderedFormKey(formKey)
-    setName(agent.name)
-    setDescription(agent.description ?? '')
-    setRole(agent.role ?? '')
-    setIconPickerOpen(false)
-    setAgentIcon(createAgentIconSelection(agent))
-  }
-
   const handleOpenChange = (nextOpen: boolean) => {
-    if (nextOpen) {
-      setName(agent.name)
-      setDescription(agent.description ?? '')
-      setRole(agent.role ?? '')
-      setAgentIcon(createAgentIconSelection(agent))
-    } else {
-      setIconPickerOpen(false)
-    }
+    if (!nextOpen) setIconPickerOpen(false)
     onOpenChange(nextOpen)
   }
 
@@ -153,11 +135,7 @@ export function EditAgentDialog({ agent, formKey, open, onOpenChange }: EditAgen
               {t(($) => $['roster.editDialog.description'])}
             </DialogDescription>
           </div>
-          <Form<AgentFormValues>
-            key={formKey}
-            className="min-h-0 flex-1"
-            onFormSubmit={handleSubmit}
-          >
+          <Form<AgentFormValues> className="min-h-0 flex-1" onFormSubmit={handleSubmit}>
             <AgentFormFields
               description={description}
               icon={agentIcon}
